feat: Auto-assign students added to the class
Added a new option to decide if the students added to the selected classes are automatically assigned to the activity.
Only those activities ending more than 24 hours ahead the student enrolment date are considered.
If your activity is always open to students, it is only considered if not manually closed.

fix: assignableInstance model had default values on some attributes, but it was incorrectly user defaultValue -> defaultTo